Move code, remove workers and fix tests
